refactor: mv terminal-theme(.spec).ts files to renderer
