+++ packages/renderer/src/lib/preferences/PreferencesConnectionDetailsTerminal.svelte
@@ -13,7 +13,7 @@ import { getExistingTerminal, registerTerminal } from '/@/stores/provider-termin
 import type { ProviderContainerConnectionInfo, ProviderInfo, ProviderVmConnectionInfo } from '/@api/provider-info';
 
 import { TerminalSettings } from '../../../../main/src/plugin/terminal-settings';
-import { getTerminalTheme } from '../../../../main/src/plugin/terminal-theme';
+import { getTerminalTheme } from '../terminal/terminal-theme';
 import NoLogIcon from '../ui/NoLogIcon.svelte';
 
 interface ProviderDetailsTerminalProps {
