+++ packages/renderer/src/lib/terminal/terminal-theme.spec.ts
