+++ packages/renderer/src/lib/terminal/terminal-theme.ts
